split up cli
