+++ src/prefect_cloud/cli/__init__.py
@@ -3,4 +3,7 @@
 # Import CLI submodules to register them to the app
 # isort: split
 
+
+from . import auth  # noqa: F401
+from . import deployments  # noqa: F401
 from . import github  # noqa: F401
